feature-policy/experimental-features/vertical-scroll-scrollintoview.tentative.sub.html

<!DOCTYPE html>
<html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/feature-policy/experimental-features/resources/vertical-scroll.js"></script>
<style>
html, body {
  height: 100%;
  width: 100%;
}

iframe {
  width: 95%;
  height: 95%;
  overflow: scroll;
  margin-top: 200%;
}

.spacer {
  width: 100%;
  height: 100%;
  margin-top: 100%;
  margin-bottom: 100%;
}

</style>
<p> An &lt;iframe&gt; further below which is not allowed to block scroll.</p>
<div class="spacer"></div>
<iframe></iframe>
<p> Making sure there is room for vertical scroll </p>
<script>
  "use strict";

  let fileName = "vertical-scroll-scrollintoview.html";
  let iframeElement = document.querySelector("iframe");

  function iframeBounds() {
    return iframeElement.getBoundingClientRect();
  }

  // Enabled 'vertical-scroll': scrollIntoView should work in all frames.
  promise_test(async() => {
    window.scrollTo(0, 0);
    await loadUrlInIframe(iframeElement, getUrl(fileName));

    await sendMessageAndGetResponse(
      iframeElement.contentWindow,
      {type: "child-frame-bounds"}).then((response) => {
        let iframeBoundsAtOrigin = {
          x: 0,
          y: 0,
          width: iframeBounds().width,
          height: iframeBounds().height};
          let nestedFrameBounds = response.bounds;
          assert_false(rects_intersect(iframeBoundsAtOrigin, nestedFrameBounds),
            "Nested <iframe> should not be visible in <iframe>." +
            `Nested <iframe>'s bounds: ${rectToString(response.bounds)} but ` +
            `<iframe> size: ${iframeBounds().width}x${iframeBounds.height}.`);
      });

    // Scroll the body of inner-most frame.
    await sendMessageAndGetResponse(iframeElement.contentWindow,
                                   {type: "scroll"});
    // The page should have scrolled vertically.
      assert_greater_than(
              window.scrollY,
              0,
             "Main frame must scroll vertically.");
    }, "Calling 'scrollIntoView()' inside a <iframe> will propagate up by" +
       " default('vertical-scroll' enabled).");

  // Disabled 'vertical-scroll': The scope of scrollIntoView is within the set
  // of disabled frames (does not propagate to main frame).
  promise_test(async() => {
    window.scrollTo(0, 0);
    iframeElement.allow = "vertical-scroll 'none';";
    await loadUrlInIframe(iframeElement, getUrl(fileName));

    await sendMessageAndGetResponse(
      iframeElement.contentWindow,
      {type: "child-frame-bounds"}).then((response) => {
      let iframeBoundsAtOrigin = {
        x: 0,
        y: 0,
        width: iframeBounds().width,
        height: iframeBounds().height};
      let nestedFrameBounds = response.bounds;
      assert_false(rects_intersect(iframeBoundsAtOrigin, nestedFrameBounds),
            "Nested <iframe> should not be visible in <iframe>." +
            `Nested <iframe>'s bounds: ${rectToString(response.bounds)} but ` +
            `<iframe> size: ${iframeBounds().width}x${iframeBounds().height}.`);
      });

    // Scroll the body of inner-most frame.
    await sendMessageAndGetResponse(iframeElement.contentWindow,
      {type: "scroll"}).then((response) => {
      // Make sure the nested <iframe> is visible.
      let nestedFrameBounds = response.bounds;
      let iframeBoundsAtOrigin = {
          x: 0,
          y: 0,
          width: iframeBounds().width,
          height: iframeBounds().height};
      // The nested <iframe> should have scrolled into visible inside its containe
      // frame.
      assert_true(rects_intersect(iframeBoundsAtOrigin, nestedFrameBounds),
          "Nested <iframe> should be visible in <iframe>." +
          `Nested <iframe>'s bounds: ${rectToString(response.bounds)}.` +
          `<iframe> size: ${iframeBounds().width}, ${iframeBounds.height}.`);
      // The page however should not have scrolled.
      assert_equals(window.scrollY, 0, "Main frame must not scroll vertically.");
    });
    }, "Calling 'scrollIntoView()' inside a <iframe> with" +
       " 'vertical-scroll none;'will not propagate upwards.");
  </script>
</html>
